refactor : 패키지 구조 리펙터링
- 객체의 용도에 따라 큰 범주로 패키지 나눔
